Move GalaxySetupData into separate TU files
