+++ UI/EncyclopediaDetailPanel.cpp
@@ -42,9 +42,10 @@
 #include "../universe/UniverseObjectVisitors.h"
 #include "../universe/Universe.h"
 #include "../universe/UnlockableItem.h"
-#include "../universe/ValueRef.h"
+#include "../Empire/Empire.h"
+#include "../Empire/EmpireManager.h"
+#include "../Empire/Government.h"
 #include "../util/Directories.h"
-#include "../util/EnumText.h"
 #include "../util/GameRules.h"
 #include "../util/i18n.h"
 #include "../util/Logger.h"
