+++ client/AI/AIClientApp.cpp
@@ -8,6 +8,7 @@
 #include "../../util/Directories.h"
 #include "../../util/i18n.h"
 #include "../util/AppInterface.h"
+#include "../util/MultiplayerCommon.h"
 #include "../../network/Message.h"
 #include "../util/Random.h"
 #include "../util/Version.h"
