+++ server/ServerApp.h
@@ -10,6 +10,7 @@
 #include "../Empire/Supply.h"
 #include "../universe/Universe.h"
 #include "../util/AppInterface.h"
+#include "../util/GalaxySetupData.h"
 #include "../util/MultiplayerCommon.h"
 #include "../util/Process.h"
