+++ universe/EnumsFwd.h
@@ -7,9 +7,6 @@ enum PlanetType : int;
 enum PlanetSize : int;
 enum PlanetEnvironment : int;
 enum MeterType : int;
-enum Shape : int;
-enum Aggression : int;
-enum GalaxySetupOption : int;
 enum EmpireAffiliationType : int;
 enum DiplomaticStatus : int;
 enum UnlockableItemType : int;
