+++ util/CMakeLists.txt
@@ -5,7 +5,7 @@ target_sources(freeorioncommon
         ${CMAKE_CURRENT_LIST_DIR}/blocking_combiner.h
         ${CMAKE_CURRENT_LIST_DIR}/CheckSums.h
         ${CMAKE_CURRENT_LIST_DIR}/Directories.h
-        ${CMAKE_CURRENT_LIST_DIR}/EnumText.h
+        ${CMAKE_CURRENT_LIST_DIR}/GalaxySetupData.h
         ${CMAKE_CURRENT_LIST_DIR}/i18n.h
         ${CMAKE_CURRENT_LIST_DIR}/Logger.h
         ${CMAKE_CURRENT_LIST_DIR}/LoggerWithOptionsDB.h
@@ -31,7 +31,7 @@ target_sources(freeorioncommon
         ${CMAKE_CURRENT_LIST_DIR}/AppInterface.cpp
         ${CMAKE_CURRENT_LIST_DIR}/CheckSums.cpp
         ${CMAKE_CURRENT_LIST_DIR}/Directories.cpp
-        ${CMAKE_CURRENT_LIST_DIR}/EnumText.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/GalaxySetupData.cpp
         ${CMAKE_CURRENT_LIST_DIR}/GameRules.cpp
         ${CMAKE_CURRENT_LIST_DIR}/i18n.cpp
         ${CMAKE_CURRENT_LIST_DIR}/Logger.cpp
