+++ util/MultiplayerCommon.cpp
@@ -40,20 +40,6 @@ namespace {
         db.Add("save.auto.exit.enabled",                    UserStringNop("OPTIONS_DB_AUTOSAVE_GAME_CLOSE"),    true);
         db.AddFlag('q', "quickstart",                       UserStringNop("OPTIONS_DB_QUICKSTART"),             false);
 
-        // Common galaxy settings
-        db.Add("setup.seed",                UserStringNop("OPTIONS_DB_GAMESETUP_SEED"),                         std::string("0"),       Validator<std::string>());
-        db.Add("setup.star.count",          UserStringNop("OPTIONS_DB_GAMESETUP_STARS"),                        150,                    RangedValidator<int>(10, 5000));
-        db.Add("setup.galaxy.shape",        UserStringNop("OPTIONS_DB_GAMESETUP_GALAXY_SHAPE"),                 DISC,                   RangedValidator<Shape>(SPIRAL_2, RANDOM));
-        db.Add("setup.galaxy.age",          UserStringNop("OPTIONS_DB_GAMESETUP_GALAXY_AGE"),                   GALAXY_SETUP_MEDIUM,    RangedValidator<GalaxySetupOption>(GALAXY_SETUP_LOW, GALAXY_SETUP_RANDOM));
-        db.Add("setup.planet.density",      UserStringNop("OPTIONS_DB_GAMESETUP_PLANET_DENSITY"),               GALAXY_SETUP_MEDIUM,    RangedValidator<GalaxySetupOption>(GALAXY_SETUP_LOW, GALAXY_SETUP_RANDOM));
-        db.Add("setup.starlane.frequency",  UserStringNop("OPTIONS_DB_GAMESETUP_STARLANE_FREQUENCY"),           GALAXY_SETUP_MEDIUM,    RangedValidator<GalaxySetupOption>(GALAXY_SETUP_LOW, GALAXY_SETUP_RANDOM));
-        db.Add("setup.specials.frequency",  UserStringNop("OPTIONS_DB_GAMESETUP_SPECIALS_FREQUENCY"),           GALAXY_SETUP_MEDIUM,    RangedValidator<GalaxySetupOption>(GALAXY_SETUP_NONE, GALAXY_SETUP_RANDOM));
-        db.Add("setup.monster.frequency",   UserStringNop("OPTIONS_DB_GAMESETUP_MONSTER_FREQUENCY"),            GALAXY_SETUP_MEDIUM,    RangedValidator<GalaxySetupOption>(GALAXY_SETUP_NONE, GALAXY_SETUP_RANDOM));
-        db.Add("setup.native.frequency",    UserStringNop("OPTIONS_DB_GAMESETUP_NATIVE_FREQUENCY"),             GALAXY_SETUP_MEDIUM,    RangedValidator<GalaxySetupOption>(GALAXY_SETUP_NONE, GALAXY_SETUP_RANDOM));
-        db.Add("setup.ai.player.count",     UserStringNop("OPTIONS_DB_GAMESETUP_NUM_AI_PLAYERS"),               6,                      RangedValidator<int>(0, IApp::MAX_AI_PLAYERS()));
-        db.Add("setup.ai.aggression",       UserStringNop("OPTIONS_DB_GAMESETUP_AI_MAX_AGGRESSION"),            MANIACAL,               RangedValidator<Aggression>(BEGINNER, MANIACAL));
-
-
         // AI Testing options-- the following options are to facilitate AI testing and do not currently have an options page widget;
         // they are intended to be changed via the command line and are not currently storable in the configuration file.
         db.Add<std::string>("ai-path",                      UserStringNop("OPTIONS_DB_AI_FOLDER_PATH"),         "python/AI",
@@ -78,134 +64,6 @@ namespace {
     }
     bool temp_bool2 = RegisterGameRules(&AddRules);
 }
-
-/////////////////////////////////////////////////////
-// GalaxySetupData
-/////////////////////////////////////////////////////
-namespace {
-    // returns number in range 0 to one less than the interger representation of
-    // enum_vals_count, determined by the random seed
-    template <typename T1>
-    int GetIdx(const T1& enum_vals_count, const std::string& seed) {
-        DebugLogger() << "hashing seed: " << seed;
-        // use probably-bad but adequate for this purpose hash function to
-        // convert seed into a hash value
-        int hash_value = 223;
-        for (size_t i = 0; i < seed.length(); ++i) {
-            //DebugLogger() << "hash value: " << hash_value << " char: " << static_cast<int>(seed[i]);
-            hash_value += (seed[i] * 61);
-            hash_value %= 191;
-        }
-        DebugLogger() << "final hash value: " << hash_value
-                      << " and returning: " << hash_value % static_cast<int>(enum_vals_count)
-                      << " from 0 to " << static_cast<int>(enum_vals_count) - 1;
-        return hash_value % static_cast<int>(enum_vals_count);
-    }
-
-    static char alphanum[] = "123456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz";
-}
-
-GalaxySetupData::GalaxySetupData() :
-    size(100),
-    shape(SPIRAL_2),
-    age(GALAXY_SETUP_MEDIUM),
-    starlane_freq(GALAXY_SETUP_MEDIUM),
-    planet_density(GALAXY_SETUP_MEDIUM),
-    specials_freq(GALAXY_SETUP_MEDIUM),
-    monster_freq(GALAXY_SETUP_MEDIUM),
-    native_freq(GALAXY_SETUP_MEDIUM),
-    ai_aggr(MANIACAL),
-    encoding_empire(ALL_EMPIRES)
-{}
-
-GalaxySetupData::GalaxySetupData(GalaxySetupData&& base) :
-    seed(std::move(base.seed)),
-    size(base.size),
-    shape(base.shape),
-    age(base.age),
-    starlane_freq(base.starlane_freq),
-    planet_density(base.planet_density),
-    specials_freq(base.specials_freq),
-    monster_freq(base.monster_freq),
-    native_freq(base.native_freq),
-    ai_aggr(base.ai_aggr),
-    game_rules(std::move(base.game_rules)),
-    game_uid(std::move(base.game_uid)),
-    encoding_empire(base.encoding_empire)
-{ SetSeed(seed); }
-
-const std::string& GalaxySetupData::GetSeed() const
-{ return seed; }
-
-int GalaxySetupData::GetSize() const
-{ return size; }
-
-Shape GalaxySetupData::GetShape() const {
-    if (shape != RANDOM)
-        return shape;
-    size_t num_shapes = static_cast<size_t>(GALAXY_SHAPES) - 1; // -1 so that RANDOM isn't counted
-    return static_cast<Shape>(GetIdx(num_shapes, seed + "shape"));
-}
-
-GalaxySetupOption GalaxySetupData::GetAge() const {
-    if (age != GALAXY_SETUP_RANDOM)
-        return age;
-    return static_cast<GalaxySetupOption>(GetIdx(3, seed + "age") + 1);       // need range 1-3 for age
-}
-
-GalaxySetupOption GalaxySetupData::GetStarlaneFreq() const {
-    if (starlane_freq != GALAXY_SETUP_RANDOM)
-        return starlane_freq;
-    return static_cast<GalaxySetupOption>(GetIdx(3, seed + "lanes") + 1);     // need range 1-3 for starlane freq
-}
-
-GalaxySetupOption GalaxySetupData::GetPlanetDensity() const {
-    if (planet_density != GALAXY_SETUP_RANDOM)
-        return planet_density;
-    return static_cast<GalaxySetupOption>(GetIdx(3, seed + "planets") + 1);   // need range 1-3 for planet density
-}
-
-GalaxySetupOption GalaxySetupData::GetSpecialsFreq() const {
-    if (specials_freq != GALAXY_SETUP_RANDOM)
-        return specials_freq;
-    return static_cast<GalaxySetupOption>(GetIdx(4, seed + "specials"));      // need range 0-3 for planet density
-}
-
-GalaxySetupOption GalaxySetupData::GetMonsterFreq() const {
-    if (monster_freq != GALAXY_SETUP_RANDOM)
-        return monster_freq;
-    return static_cast<GalaxySetupOption>(GetIdx(4, seed + "monsters"));      // need range 0-3 for monster frequency
-}
-
-GalaxySetupOption GalaxySetupData::GetNativeFreq() const {
-    if (native_freq != GALAXY_SETUP_RANDOM)
-        return native_freq;
-    return static_cast<GalaxySetupOption>(GetIdx(4, seed + "natives"));       // need range 0-3 for native frequency
-}
-
-Aggression GalaxySetupData::GetAggression() const
-{ return ai_aggr; }
-
-const std::map<std::string, std::string>& GalaxySetupData::GetGameRules() const
-{ return game_rules; }
-
-const std::string& GalaxySetupData::GetGameUID() const
-{ return game_uid; }
-
-void GalaxySetupData::SetSeed(const std::string& seed_) {
-    std::string new_seed = seed_;
-    if (new_seed.empty() || new_seed == "RANDOM") {
-        ClockSeed();
-        new_seed.clear();
-        for (int i = 0; i < 8; ++i)
-            new_seed += alphanum[RandInt(0, (sizeof(alphanum) - 2))];
-        DebugLogger() << "Set empty or requested random seed to " << new_seed;
-    }
-    seed = std::move(new_seed);
-}
-
-void GalaxySetupData::SetGameUID(const std::string& game_uid_)
-{ game_uid = game_uid_; }
 
 /////////////////////////////////////////////////////
 // PlayerSetupData
